implement microservice architecture with spring cloud

+++ camp/src/main/java/ch/bfh/autumn/heroes/model/Hero.java
@@ -13,5 +13,5 @@ public class Hero {
   private String name;
   private int atk;
   private int def;
-  private int hp;
+  private double hp;
 }
